Crop the image
I hesitated to crop the original, but because it's in the public
domain, I think it's okay to do. The problem with leaving the original
text on the image is that it was basically unreadable at the small
scale I used, and looked more like a mistake than anything else.