Significant overhaul of Makefile-main.mk
1) Now uses an external config file in YAML format.

2) New approach does away with separate output file. The index.html
file is put at the top level of the user's notebook directory, and the
contents directory is also the place where the html putput files are
placed.  The idea now is that the user makes the whole directory be
the thing that is the room of the web site.  This is easily done with
GitHub Pages, and simplifies a number of things for the user and in
the code.

3) Removed unused variables like toc-args and index-args

4) Changed explicit 'create-dirs' target to use conventional make
dependencies to create directories.