1) Removed contents directory variable; 2) shuffled order in file
Decided to hardwire the name of the contents directory, so that
`setup` can be used to create initial content for a notebook.  I
couldn't think of a way to let the user set the name of the contents
directory while still using `setup` to create initial content (because
`setup` is run very early in the installation process), and I judged
that having `setup` create the initial content was more important than
allowing the contents directory to be renamed.

Allow `setup` to create the initial content will make explanations of
Pangolin Notebook much better