Change previous "$contents" variable to "%TOC%"
The previous use of a single dollar sign always confused me because
Pandoc uses two dollar signs around variables.  But the real problem
was that pandoc variables can only be used in templates, and so this
used a limited one-off subsitution trick using the command
`envsubst`.  The new version uses a more general substitution approach
employing `sed`, and can substitute more variales in the body of the
user's text.